==> FinalSubmission/EDA & Preprocessing & Labeling/STEP3_Labelling_Posts_Final_dataset.ipynb <==
# Imports

import re
import pandas as pd
import numpy as np


df_posts = pd.read_csv("df_posts_cleaned.csv")


df_labeled_comments = pd.read_csv('df_labeled_comments.csv')

label_counts = df_labeled_comments['label'].value_counts()
label_counts

# Combining comments with posts

df_comment = df_labeled_comments

df_posts = df_posts.drop(columns = ['permalink'])

df_comment

#lables
df_labeles = df_comment.groupby('submission_id')['label'].apply(list).reset_index()

from collections import Counter

def choose_label(labels, threshold = 25):
    counts = Counter(labels)

    if counts['NAH'] >= threshold:
        return 'NAH'
    elif counts['ESH'] >= threshold:
        return 'ESH'
    else:
        return counts.most_common(1)[0][0]  # most common label



df_labeles['ChosenLabel'] = df_labeles['label'].apply(choose_label)

df_labeles = df_labeles.drop(columns = 'label')
df_labeles = df_labeles.rename(columns={'ChosenLabel': 'Label'})

# Assigning lables to posts

df_labeles.shape

df_posts.shape

labeled_post_ids = df_labeles['submission_id'].unique()

# dropping posts that wouln't get lables assigned
df_posts = df_posts[df_posts['submission_id'].isin(labeled_post_ids)]


df_posts.shape

labeled_df = pd.merge(df_posts, df_labeles, on='submission_id', how='inner')

labeled_df.to_csv('labeled_df')


labeled_df['Label'].value_counts()

# Creating Final Subset:
This subset will hold only 500 records of minority classes only (YTA, ESH, NAH)




minorities = labeled_df[labeled_df['Label'] != 'NTA']
minorities_subset = minorities[:500]

minorities_subset.to_csv('AITA_minorities_subset.csv')
